Slash commands, partial config to env var switch
Seemed to not be an issue with the bridge commands as I thought, but rather wrong type hint for discord.py. Oh well, keeping it all slash commands if that is the way of the future.
https://github.com/Pycord-Development/pycord/issues/1331

Also have started going through and switching from the config.yml file to env vars. Initially planned on doing that afterward, but may as well work on it now if I'm rewriting things.

Also have begun going through and redoing readme with that change in mind.